feat: comment support for redis
