+++ Makefile
@@ -47,7 +47,7 @@ test:
 
 .PHONY: run
 run:
-	go run main.go
+	go run cmd/main.go
 
 .PHONY: build
 build:
